Move out create request queue, unit test
