+++ src/ray/object_manager/plasma/connection.h
@@ -16,8 +16,13 @@ class Client;
 using PlasmaStoreMessageHandler = 
     std::function<ray::Status(std::shared_ptr<Client>, flatbuf::MessageType, const std::vector<uint8_t>&)>;
 
+class ClientInterface {
+ public:
+  virtual ~ClientInterface() {}
+};
+
 /// Contains all information that is associated with a Plasma store client.
-class Client : public ray::ClientConnection  {
+class Client : public ray::ClientConnection, public ClientInterface  {
  public:
   static std::shared_ptr<Client> Create(
       PlasmaStoreMessageHandler message_handler, ray::local_stream_socket &&socket);
