+++ src/ray/object_manager/test/create_request_queue_test.cc
@@ -1,0 +1,54 @@
+// Copyright 2017 The Ray Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "ray/object_manager/plasma/create_request_queue.h"
+
+#include "gtest/gtest.h"
+#include "ray/common/status.h"
+
+namespace plasma {
+
+class MockClient : public ClientInterface {
+ public:
+  MockClient() {}
+};
+
+TEST(CreateRequestQueueTest, TestSimple) {
+  int num_store_full;
+  CreateRequestQueue queue(
+      /*max_retries=*/0,
+      /*evict_if_full=*/true,
+      /*on_store_full=*/[&]() {
+      num_store_full++;
+      }
+      );
+
+  bool created = false;
+  auto request = [&](bool reply_on_oom, bool evict_if_full) {
+    created = true;
+    return Status();
+  };
+  auto client = std::make_shared<MockClient>();
+  queue.AddRequest(client, request);
+  ASSERT_FALSE(created);
+  ASSERT_TRUE(queue.ProcessRequests().ok());
+  ASSERT_TRUE(created);
+}
+
+}  // namespace plasma
+
+int main(int argc, char **argv) {
+  ::testing::InitGoogleTest(&argc, argv);
+  return RUN_ALL_TESTS();
+}
